requirements.txt
dnslib==0.9.23
dnspython==2.4.2
matplotlib==3.8.2 